Update the multi service call to better handle multiple service calls
Using an operation queue alongside a dispatch group ensures that all service calls are completed before letter the model know that all the data has been retrieved.
The dispatchGroup.leave() has also been relocated to the completion handlers for the service calls as the .leave() was being called prematurely, causing some data to be loaded and others not.